added changes for org dataset crud operation v2

+++ src/main/java/com/winwin/winwin/controller/OrganizationDataSetController.java
@@ -6,6 +6,7 @@ import javax.transaction.Transactional;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -15,6 +16,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.winwin.winwin.payload.OrganizationDataSetPayLoad;
 import com.winwin.winwin.service.OrganizationDataSetService;
 import com.winwin.winwin.entity.OrganizationDataSet;
+import com.winwin.winwin.exception.OrganizationDataSetException;
 
 /**
  * @author ArvindK
@@ -22,37 +24,67 @@ import com.winwin.winwin.entity.OrganizationDataSet;
  */
 @RestController
 @RequestMapping(value = "/organizationdataset")
-public class OrganizationDataSetController {
+public class OrganizationDataSetController extends BaseController {
 
 	@Autowired
 	private OrganizationDataSetService organizationDataSetService;
 
-	@GetMapping("/organizationdataset")
-	public String getOrganizationDataSet() {
-		return "Hello World  For Organization Data Set";
-	}
-
+	@SuppressWarnings("rawtypes")
 	@PostMapping("/create")
 	@Transactional
-	public void createOrganizationDataSet(@Valid @RequestBody OrganizationDataSetPayLoad organizationDataSetPayLoad) {
-		organizationDataSetService.createOrUpdateOrganizationDataSet(organizationDataSetPayLoad);
-	}
-	
-	@PostMapping("/update")
-	@Transactional
-	public void updateOrganizationDataSet(@Valid @RequestBody OrganizationDataSetPayLoad organizationDataSetPayLoad) {
-		organizationDataSetService.createOrUpdateOrganizationDataSet(organizationDataSetPayLoad);
-	}
-	
-	@PostMapping("/delete")
-	@Transactional
-	public void deleteOrganizationDataSet(@Valid @RequestBody Long id) {
-		organizationDataSetService.removeOrganizationDataSet(id);
+	public ResponseEntity createOrganizationDataSet(
+			@Valid @RequestBody OrganizationDataSetPayLoad organizationDataSetPayLoad)
+			throws OrganizationDataSetException {
+		try {
+			organizationDataSetService.createOrUpdateOrganizationDataSet(organizationDataSetPayLoad);
+		} catch (Exception e) {
+			throw new OrganizationDataSetException("error occoured while creating organization dataset");
+		}
+		return sendSuccessResponse("org.success.created");
 	}
 
-	@GetMapping("/organizationdatasetlist")
-	public List<OrganizationDataSet> getOrganizationDataSetList() {
-		return organizationDataSetService.getOrganizationDataSetList();
-	}// end of method getOrganizationDataSetList
+	@SuppressWarnings("rawtypes")
+	@PostMapping("/update")
+	@Transactional
+	public ResponseEntity updateOrganizationDataSet(
+			@Valid @RequestBody OrganizationDataSetPayLoad organizationDataSetPayLoad)
+			throws OrganizationDataSetException {
+		try {
+			organizationDataSetService.createOrUpdateOrganizationDataSet(organizationDataSetPayLoad);
+		} catch (Exception e) {
+			throw new OrganizationDataSetException("error occoured while updating organization dataset");
+		}
+		return sendSuccessResponse("org.success.updated");
+	}
+
+	@SuppressWarnings("rawtypes")
+	@PostMapping("/delete")
+	@Transactional
+	public ResponseEntity deleteOrganizationDataSet(
+			@Valid @RequestBody OrganizationDataSetPayLoad organizationDataSetPayLoad)
+			throws OrganizationDataSetException {
+		try {
+			if (null != organizationDataSetPayLoad) {
+				Long id = organizationDataSetPayLoad.getId();
+				organizationDataSetService.removeOrganizationDataSet(id);
+			}
+
+		} catch (Exception e) {
+			throw new OrganizationDataSetException("error occoured while deleting organization dataset");
+		}
+		return sendSuccessResponse("org.success.deleted");
+	}
+
+	@GetMapping("/list")
+	public ResponseEntity<?> getOrganizationDataSetList() throws OrganizationDataSetException {
+		List<OrganizationDataSet> orgDataSetList = null;
+		try {
+			orgDataSetList = organizationDataSetService.getOrganizationDataSetList();
+		} catch (Exception e) {
+			throw new OrganizationDataSetException("error occoured while gettting organization dataset list");
+		}
+		return sendSuccessResponse(orgDataSetList);
+
+	}
 
 }
